backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary>=2.9.1
djangorestframework==3.15.1
channels==4.0.0
channels_redis==4.0.0
daphne==4.0.0
django-cors-headers==4.4.0
redis>=4.2.0
